Make api folder changes, add new endpoints
